+++ packages/suite/src/storage/migrations/index.ts
@@ -752,4 +752,14 @@ export const migrate: OnUpgradeFunc<SuiteDBSchema> = async (
             return metadata;
         });
     }
+
+    if (oldVersion < 44) {
+        await updateAll(transaction, 'txs', tx => {
+            if (tx.tx.tokens.length > 0 && tx.tx.rates) {
+                delete tx.tx.rates;
+            }
+
+            return tx;
+        });
+    }
 };